refactor: using the pages instead of app
