--- .backup/README.md
@@ -1,2 +1,0 @@
-# my_website
-My personal website's source code
