--- .backup/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "lucapalmi.com",
-  "version": "1.0.0",
-  "description": "lucapalmi.com website",
-  "main": "index.js",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Lucino772/lucapalmi.com.git"
-  },
-  "author": "Lucino772",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Lucino772/lucapalmi.com/issues"
-  },
-  "homepage": "https://github.com/Lucino772/lucapalmi.com#readme",
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "framer-motion": "^4.1.17",
-    "next": "^10.2.3",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.2",
-    "sass": "^1.34.1",
-    "styled-components": "^5.3.0"
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.3.7",
-    "postcss": "^8.3.11",
-    "tailwindcss": "^2.2.17"
-  }
-}
