--- .backup/pages/_app.js
@@ -1,11 +1,0 @@
-import "../public/base.scss";
-
-import { AnimatePresence } from "framer-motion";
-
-export default function App({Component, pageProps, router}) {
-    return (
-        <AnimatePresence exitBeforeEnter>
-            <Component {...pageProps} key={router.route} />
-        </AnimatePresence>
-    )
-}
