+++ src/pages/about.tsx
@@ -1,19 +1,16 @@
 "use client"
-// import Image from 'next/image'
-// import { Inter } from '@next/font/google'
-// import styles from './page.module.css'
 
 import Logo from "../components/Logo";
 import NavBar from "../components/NavBar";
-import Welcome from "../components/Welcome";
+import AboutText from "../components/AboutText";
 
-export default function Index() {
+export default function About() {
     return (
         <>
             <NavBar showLogo={false} exitDelay={0} />
             <div className="absolute w-full left-0 right-0 top-[50%] -translate-y-1/2 flex justify-evenly items-center overflow-hidden">
                 <Logo width={312} height={312} />
-                <Welcome />
+                <AboutText />
             </div>
         </>
     )
